fix overlaping components in product page by adding z index to header
