+++ client-frontend/src/components/organisms/CheckoutPageComponent.jsx
@@ -7,6 +7,7 @@ import ShippingAddressesComponent from "../moleculas/ShippingAddressesComponent"
 import {submitOrder} from "../../api/OrderApi";
 import PopupSuccessAlert from "../atoms/alerts/PopupSuccessAlert";
 import DangerAlert from "../atoms/alerts/DangerAlert";
+import {getBuyerAddresses} from "../../api/BuyerApi";
 
 function CheckoutPageComponent(){
 
@@ -22,7 +23,8 @@ function CheckoutPageComponent(){
         }
     })
 
-    const [shippingAddress, setShippingAddress] = useState(null)
+    const [shippingAddresses, setShippingAddresses]= useState([])
+    const [selectedShippingAddress, setSelectedShippingAddress] = useState(null)
 
     const {username} = useAuth()
     const location = useLocation()
@@ -46,16 +48,26 @@ function CheckoutPageComponent(){
             )
     }
 
-    function callback(shippingAddress){
-        if(!!shippingAddress){
-            setDangerAlertShowing(false)
-        }
-        setShippingAddress(shippingAddress)
+    function getShippingAddresses(){
+        getBuyerAddresses()
+            .then(
+                (response) => {
+                    setShippingAddresses(response.data)
+                    setSelectedShippingAddress(response.data[0])
+                }
+            )
+            .catch(
+                (err) => console.log(err)
+            )
+    }
+
+    function handleAddressSelected(shippingAddress){
+        setSelectedShippingAddress(shippingAddress)
     }
 
     function handlePlaceOrder(){
-        if(!!shippingAddress){
-            submitOrder(shippingAddress,checkoutItems,username)
+        if(!!selectedShippingAddress){
+            submitOrder(selectedShippingAddress,checkoutItems,username)
                 .then(
                     () => {
                         setInfoAlertShowing(true)
@@ -80,15 +92,17 @@ function CheckoutPageComponent(){
     useEffect(() => {
         if(username){
             getCartItemsList()
+            getShippingAddresses()
         }
     }, [location, username]);
 
     return (
+        // todo type for alert
         <div className="mb-96">
             {!!infoAlertShowing &&
                 <PopupSuccessAlert classname="top-[86px] right-5 sm:top-2 sm:mt-4" handleCloseButton={popupInfoCloseButton} title="Order Placed" paragraph="You will be redirected..."/>
             }
-            {!!dangerAlertShowing &&
+            {!dangerAlertShowing &&
                 <DangerAlert className="top-[86px] right-5 sm:top-2 sm:mt-4" paragraph="To place an order you nedd to selet a shipping address or add a new address"/>
             }
 
@@ -98,7 +112,11 @@ function CheckoutPageComponent(){
                     <Link to="/account/cart" className="text-sm font-semibold leading-6 text-inherit dark:text-inherit">
                         <span aria-hidden="true">&larr;</span> Cart
                     </Link>
-                    <ShippingAddressesComponent className="mt-4" onClick={callback}/>
+                    <div className="mt-4">
+                        <ShippingAddressesComponent shippingAddresses={shippingAddresses}
+                                                    selectedShippingAddress={selectedShippingAddress}
+                                                    onAddressSelected={handleAddressSelected}/>
+                    </div>
                 </div>
 
                 <div className="w-1/2 max-w-lg sm:w-full sm:mx-auto sm:mt-10">
